requirements.txt
aiohttp>=3.8.0
tqdm>=4.65.0
mysql-connector-python>=8.0.0
certifi>=2023.7.22